+++ src/Pim/Bundle/CatalogBundle/Model/AttributeInterface.php
@@ -105,9 +105,19 @@ interface AttributeInterface extends TimestampableInterface, TranslatableInterfa
 
     /**
      * Get available locales
+     *
+     * @deprecated will be removed in 1.4
+     *
      * @return ArrayCollection|null
      */
     public function getAvailableLocales();
+
+    /**
+     * Get available locale codes
+     *
+     * @return array|null
+     */
+    public function getAvailableLocaleCodes();
 
     /**
      * Set entity type