+++ CHANGELOG.md
@@ -63,11 +63,15 @@
 - Remove the `Pim\Bundle\CatalogBundle\Doctrine\(ORM|MongoDBODM)\Filter\BaseFilter` to use proper dedicated filters
 - The parameter `category_id` for the route `pim_enrich_product_listcategories` has been renamed to `categoryId`
 - Change constructor of `Pim\Bundle\BaseConnectorBundle\Reader\File\CsvProductReader`. Now `FieldNameBuilder`, channel, locale and currency entity classes are mandatory.
+- Add getAvailableLocaleCodes in AttributeInterface
 
 ## Bug fixes
 - PIM-3332: Fix incompatibility with overriden category due to usage of ParamConverter in ProductController
 
 # 1.2.x
+
+## Bug fixes
+- PIM-3298 : Fix issue with locale specific property of an attribute when edit and mass edit
 
 # 1.2.11 (2014-10-31)